finalise the logic for user subnitting a guess, checking if user exists, match exists, user hasnt played it before and then storing the usergame, then returning a response  with the results of the guess, created custom query for checking usergame by userid and matchid, fixed the submit guess endpoint controller to accept a request body, manually inserted a test user in the users table to enable testing,

+++ src/main/java/com/retroscore/repository/MatchRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface MatchRepository extends JpaRepository<Match,Integer> {
+public interface MatchRepository extends JpaRepository<Match,Long> {
 
     @Query("SELECT m FROM Match m WHERE m.homeTeam.id = :teamId OR m.awayTeam.id= :teamId")
     List<Match> findByTeamId(@Param("teamId") Long teamId);
